Refactor bitmap and pixel to use struct instead of pointer.

+++ includes/pixel.h
@@ -1,13 +1,11 @@
-#include <stdint.h>
-
 #ifndef _PIXEL_H_
 #define _PIXEL_H_
 
 typedef struct pixel {
-    uint8_t r;
-    uint8_t g;
-    uint8_t b;
-    uint8_t alpha;
+    int r;
+    int g;
+    int b;
+    int alpha;
 } Pixel_t;
 
 extern Pixel_t Pixel_new();
